Initial code commit
With this commit, you can run the server and the client (run the server first). The client will send input to the server, the server will process that input and send it back to the client, and then the client will update its game state. Both the client and the server launch a window to see the current state of the game from each perspective.

There are several TODOs in the code detailing ways to improve the code.