Add artificial network latency
The server sets the artificial network latency (network delay and jitter) and communicates these settings to the client via the new "settings channel" (which is TCP based).

Pressing the D or J keys (or Shift+D or Shift+J) *in the server window* will increase (decrease) these network latency settings for both the server and the client.